Feature: api.py: GET price configuration data
Add a new public method to the plenty_api module, that fetches
configurations for different sales prices setup in Plentymarkets.
Implement it together with a minimal parameter, that shrinks the output
to a minimum.
Change the implementation of get_route to utilize a mapping table, which
is defined in plenty_api/valid_parameters.py.
Make the imported module name 'param' more clear by changing it to
constants.
Add reference URLs to the api.py class description for a couple of
public methods.

Add tests and documentation for the new method.

Signed-off-by: Sebastian Fricke <sebastian.fricke.linux@gmail.com>