Refactor: api.py: print request url for debugging
print out the URL, which was used for the request in order to get more
direct feedback on errors.
It is not clear to me currently if the other 2 print functions actually
add any direct value, as they print a trimmed down version of the
request URL. The only use-case is to show a malfunction of
build_endpoint.

Signed-off-by: Sebastian Fricke <sebastian.fricke.linux@gmail.com>