Bug-Fix: get_orders_by_date: Missing elements
Fix a bug, which caused additional elements declared in the request url
to be missing from the response body. The reason for this was the
special case handling of order requests.
Most requests at the PlentyMarkets API have the following syntax for
handling additonal elements: ...com/rest/items?with={element1},{element2},
but orders have the following format:
...com/rest/items?with%5B%5D={element1}&with%5B%5D={element2}

The previous version was missing the %5B%5D ([]) and this case slipt
through my testing.

Signed-off-by: Sebastian Fricke <sebastian.fricke.linux@gmail.com>