Feature: api.py: GET referrers
Implement the referrer get request, to fetch all order referrers from
plentymarkets. This request acts a little bit different, thus a small
adjustment in the plenty_api_request method had to be made.
Write the documentation for the public method into the class description
as well as the API reference.

Make a small improvement of the module description.

Signed-off-by: Sebastian Fricke <sebastian.fricke.linux@gmail.com>