utils.py: Check if required JSON fields are filled
Modify the `sanity_check_json` function, to not only check if a required
field exists, but also if it contains atleast one value. Adjust the
constant definitions in constants.py, to declare the expected type.
Add the helper function `json_field_filled`, which is used to check a
single field for a valid value.

Signed-off-by: Sebastian Fricke <sebastian.fricke@posteo.net>