api.py: Add new routes for the creation of redistributions
The complete creation of an redistribution contains multiple actions
like creating the order, filling it with transactions, setting event
dates and booking stock movements. The `plenty_api_create_redistribution`
method tries to provide an abstraction upon that process, that allows
the user to perform multiple actions at once.
In order to do so the method reads out a template data structure, that
contains a simplified version of the required JSON for the REST API,
together with optional elements that describe incoming or outgoing
transactions. The user can then choose if those transactions are booked
automatically or later manually by the user.

This commit contains 4 new public methods for:
- the creation of redistributions
- the creation of transactions
- booking all transactions of an order
- updating specific attributes of a redistribution

Signed-off-by: Sebastian Fricke <sebastian.fricke@posteo.net>