api.py: send error information to public methods
Do not handle all errors in `__plenty_api_request`, return the error
JSON response and let the publich method indiviually decide how to
handle the error. Add error prints to all existing GET routes.

For now we keep returning `None` on error, until we decide upon a clean
and smart way to send error responses to the user. The current behavior
enables to simply check for None with the response of a plenty api
method, so don't break that behavior.

Signed-off-by: Sebastian Fricke <sebastian.fricke@posteo.net>