Feature: CLI progress bar
Add the option to activate a tqdm progress bar for GET requests
with multiple pages. The option can be toggled by setting the public
class attribute `cli_progress_bar` to either True or False.
Add a documentation entry with an example, update the pyproject.toml
file and the poetry.lock file.

Signed-off-by: Sebastian Fricke <sebastian.fricke@posteo.net>